Added features
Sidebar
Labeled sidebar and changed selectbox to radio selection.
Added “about” section

Information page:
Used info.md as a guideline for the user on how to use the app. To convert the markdown file to text, “markdown” and “bs4” libraries have to be imported first
 Added table for sentiment descriptions

Added EDA and insights pages

EDA
Added options to preview the head or tail of the dataframe under the “preview dataframe” checkbox
Added a dropdown to view visualisations by sentiment, for instance 'Distribution of Tweet Word Count' or description of the word count column or hashtag plots